[core] Split GeoJSONTileData and related classes into a separate header
